+++ src/modules/Elsa.Workflows.Core/Contracts/IActivityStateProtector.cs
@@ -1,0 +1,10 @@
+// using System.Text.Json;
+//
+// namespace Elsa.Workflows;
+//
+// public interface IActivityStateProtector
+// {
+//     string Encode(ProtectedActivityStateContext context);
+// }
+//
+// public record ProtectedActivityStateContext(ActivityExecutionContext ActivityExecutionContext, InputDes);
